Factor choice of rand out of Distribution
It can be useful for the programmer to be able to specify a seed.
This isn't possible if ThreadLocalRandom is used as a private val
inside of Distribution. I factored it out by parametrizing Distribution
over rand and providing a default that is equivalent to the original.

The only incompatilibity this should introduce is with the change of
freeze. Because it creates a distribution (i.e., uses rand) its
natural place seems to be in the new Distributions class.
Otherwise a programmer could accidentally introduce seemingly
nondeterministic behavior even with a preset seed.